Finished first working version
The webpage's logic was completed and bundled using WebPack. The JS tab switching logic's structure was done using modules and the module pattern to create the tab switching. The style of the page is barebones and will be improved upon in an upcoming change. The current goal is to optimize the code and maybe add some QoL features.